requirements.txt
pdfrw>=0.4
pyyaml>=6.0
